Don't condition behavior on existence of attributes.
The new `codesearch-py` library has typesafe and always-existing
attributes, which is an API change from the former model of having
attributes only exist if they also existed in the server response.

Type safety is great. This change updates some of the interface code to
deal with the new way of interacting with server responses.